Final
This is the Final form of the project. 
I removed the Weighted moving average and Exponential forecasting features as they can be implemented with the same ways I implemented the other forecasting features and adding them would be a waste of time and I would rather use that time doing other projects and learning new technologies.
The sizes of the windows are also kind of broken and dont size properly as I can't be bothered to fix it honestly as I am excited to start on a new project.

Let me recap everything this app can do:
This app can forecast 2022 Sales data using Naive, and Moving Average Forecasting, the user inputs data for 2021 and 2022, they can do that by manually adding every month's data, using a CSV file, or skipping the process and generating random data for convenience.